requirements.txt
pandas
selenium
pyautogui
ttkbootstrap
openpyxl
python-docx
